refactor: 検索処理を別クラスに切り分けた
検索処理をするクラスが検索した日時を持つようにすれば自然になる
